refactor(poster): Seperate getPoster functions for GEWIS/Hubble
