+++ src/handlers/poster/hubble/index.tsx
@@ -1,7 +1,14 @@
 import './components/index.scss';
-import PosterBaseView from '../index.tsx';
+import PosterBaseView from '../';
+import { getGewisPosters } from '../../api';
 import OverlayHubble from './components/OverlayHubble.tsx';
 
 export default function PosterHubbleView() {
-  return <PosterBaseView overlay={OverlayHubble} />;
+  return (
+    <PosterBaseView
+      overlay={OverlayHubble}
+      // @ts-expect-error This should be type converting
+      getPosters={getGewisPosters}
+    />
+  );
 }
